+++ packages/eyes-cypress-universal/test/fixtures/testApp/cypress/integration-play/play.js
@@ -9,9 +9,10 @@ describe('Play Cypress', () => {
     cy.eyesOpen({
       appName: 'Play Cypress',
     });
-    cy.eyesCheckWindow({
-      tag: 'Play Cypress',
-    });
+    // cy.eyesCheckWindow({
+    //   tag: 'Play Cypress',
+    //   layout: [{selector: 'body > div > h1'},   {top: 100, left: 0, width: 1000, height: 100},]
+    // });
     cy.eyesCheckWindow({
       target: 'region',
       selector: {
@@ -19,6 +20,39 @@ describe('Play Cypress', () => {
         selector: 'body > div > h1' 
       }
     });
+    cy.eyesClose();
+    cy.eyesGetAllTestResults().then((results) => {
+      console.log(results)
+    })
+  });
+  it.only('test region in shadow DOM', () => {
+    cy.visit('https://applitools.github.io/demo/TestPages/ShadowDOM/index.html');
+    cy.eyesOpen({
+      appName: 'some app',
+      testName: 'region in shadow dom',
+      browser: {width: 800, height: 600},
+    });
+
+    // cy.eyesCheckWindow({
+    //   target: 'region',
+    //   selector: {
+    //     type: 'css',
+    //     selector: '#has-shadow-root'
+    //   }
+    // });
+    cy.eyesCheckWindow({
+      target: 'region',
+      selector: [{
+        type: 'css',
+        selector: '#has-shadow-root',
+        nodeType: 'shadow-root'
+      },{
+          type: 'css',
+          selector: 'h1',
+          nodeType: 'element'
+          
+      }]
+  });
     cy.eyesClose();
   });
 });