+++ packages/eyes-cypress-universal/test/fixtures/testApp/cypress/integration-run/region-in-shadow-dom.js
@@ -32,28 +32,28 @@ describe('Region in shadow DOM', () => {
     cy.eyesClose();
   });
 
-  it('test region in shadow DOM', () => {
-    cy.visit('https://applitools.github.io/demo/TestPages/ShadowDOM/index.html');
-    cy.eyesOpen({
-      appName: 'some app',
-      testName: 'region in shadow dom',
-      browser: {width: 800, height: 600},
-    });
-    cy.eyesCheckWindow({
-      target: 'region',
-      selector: [
-        {
-          type: 'css',
-          selector: '#has-shadow-root',
-          nodeType: 'shadow-root',
-        },
-        {
-          type: 'css',
-          selector: 'h1',
-          nodeType: 'element',
-        },
-      ],
-    });
-    cy.eyesClose();
+it.only('test region in shadow DOM', () => {
+  cy.visit('https://applitools.github.io/demo/TestPages/ShadowDOM/index.html');
+  cy.eyesOpen({
+    appName: 'some app',
+    testName: 'region in shadow dom',
+    browser: {width: 800, height: 600},
   });
+  cy.eyesCheckWindow({
+    target: 'region',
+    selector: [
+      {
+        type: 'css',
+        selector: '#has-shadow-root',
+        nodeType: 'shadow-root',
+      },
+      {
+        type: 'css',
+        selector: 'h1',
+        nodeType: 'element',
+      },
+    ],
+  });
+  cy.eyesClose();
+});
 });